Change from float to double

+++ src/common/cli.h
@@ -65,6 +65,6 @@ typedef struct
 cli_bps_t;
 
 extern void cli_display(cli_t *) __attribute__((nonnull(1)));
-extern float cli_calc_bps(cli_bps_t *) __attribute__((nonnull(1)));
+extern double cli_calc_bps(cli_bps_t *) __attribute__((nonnull(1)));
 
 #endif /* _COMMON_CLI_H_ */
